Added All CORE Functionalities
Need to touch some UI AND add few additional functionalities

+++ After finishing 2nd year (todo list)/ToDo app/app/src/main/java/com/example/tasktodo/DBHandler.java
@@ -80,7 +80,7 @@ public class DBHandler extends SQLiteOpenHelper {
         db.close();
     }
 
-    public void updateTask(int id, Task task) {
+    public void updateTask(long id, Task task) {
         SQLiteDatabase db = this.getWritableDatabase();
         ContentValues values = new ContentValues();
